requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.2
cryptography==36.0.0
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
easy-thumbnails==2.8
freezegun==1.1.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
libmagic==1.0
lxml==4.6.4
MarkupSafe==2.0.1
mock==4.0.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-magic==0.4.24
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
shortuuid==1.0.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
svglib==1.1.0
tinycss2==1.1.0
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
